move the supergraph service to a subfolder (#4204)
no functional change, this move `supergraph_service.rs` in
`src/services/supergraph/` and splits the tests in a separate file
