+++ apollo-router/src/services/supergraph/tests.rs
@@ -1,866 +1,24 @@
-//! Implements the router phase of the request lifecycle.
-
-use std::sync::atomic::Ordering;
+use std::collections::HashMap;
 use std::sync::Arc;
-use std::task::Poll;
-use std::time::Instant;
+use std::time::Duration;
 
-use futures::future::BoxFuture;
-use futures::stream::StreamExt;
-use futures::TryFutureExt;
-use http::StatusCode;
-use indexmap::IndexMap;
-use router_bridge::planner::Planner;
-use router_bridge::planner::UsageReporting;
-use tokio::sync::mpsc;
-use tokio::sync::mpsc::error::SendError;
-use tokio_stream::wrappers::ReceiverStream;
-use tower::BoxError;
-use tower::Layer;
-use tower::ServiceBuilder;
+use http::HeaderValue;
 use tower::ServiceExt;
 use tower_service::Service;
-use tracing::field;
-use tracing::Span;
-use tracing_futures::Instrument;
 
-use super::execution::QueryPlan;
-use super::layers::allow_only_http_post_mutations::AllowOnlyHttpPostMutationsLayer;
-use super::layers::content_negotiation;
-use super::layers::persisted_queries::PersistedQueryLayer;
-use super::layers::query_analysis::ParsedDocument;
-use super::layers::query_analysis::QueryAnalysisLayer;
-use super::new_service::ServiceFactory;
-use super::router::ClientRequestAccepts;
-use super::subgraph_service::MakeSubgraphService;
-use super::subgraph_service::SubgraphServiceFactory;
-use super::ExecutionServiceFactory;
-use super::QueryPlannerContent;
-use crate::configuration::Batching;
-use crate::context::OPERATION_NAME;
-use crate::error::CacheResolverError;
 use crate::graphql;
-use crate::graphql::IntoGraphQLErrors;
-use crate::graphql::Response;
-use crate::notification::HandleStream;
-use crate::plugin::DynPlugin;
-use crate::plugins::subscription::SubscriptionConfig;
-use crate::plugins::telemetry::tracing::apollo_telemetry::APOLLO_PRIVATE_DURATION_NS;
-use crate::plugins::telemetry::Telemetry;
-use crate::plugins::telemetry::LOGGING_DISPLAY_BODY;
-use crate::plugins::traffic_shaping::TrafficShaping;
-use crate::plugins::traffic_shaping::APOLLO_TRAFFIC_SHAPING;
-use crate::query_planner::subscription::SubscriptionHandle;
-use crate::query_planner::subscription::OPENED_SUBSCRIPTIONS;
-use crate::query_planner::subscription::SUBSCRIPTION_EVENT_SPAN_NAME;
-use crate::query_planner::BridgeQueryPlanner;
-use crate::query_planner::CachingQueryPlanner;
-use crate::query_planner::QueryPlanResult;
-use crate::query_planner::WarmUpCachingQueryKey;
-use crate::router_factory::create_plugins;
-use crate::router_factory::create_subgraph_services;
-use crate::services::query_planner;
+use crate::plugin::test::MockSubgraph;
+use crate::services::router::ClientRequestAccepts;
+use crate::services::subgraph;
 use crate::services::supergraph;
-use crate::services::ExecutionRequest;
-use crate::services::ExecutionResponse;
-use crate::services::QueryPlannerResponse;
-use crate::services::SupergraphRequest;
-use crate::services::SupergraphResponse;
-use crate::spec::Query;
 use crate::spec::Schema;
+use crate::test_harness::MockedSubgraphs;
 use crate::Configuration;
 use crate::Context;
 use crate::Notify;
+use crate::TestHarness;
 
-pub(crate) const QUERY_PLANNING_SPAN_NAME: &str = "query_planning";
-
-/// An [`IndexMap`] of available plugins.
-pub(crate) type Plugins = IndexMap<String, Box<dyn DynPlugin>>;
-
-/// Containing [`Service`] in the request lifecyle.
-#[derive(Clone)]
-pub(crate) struct SupergraphService {
-    execution_service_factory: ExecutionServiceFactory,
-    query_planner_service: CachingQueryPlanner<BridgeQueryPlanner>,
-    schema: Arc<Schema>,
-    notify: Notify<String, graphql::Response>,
-}
-
-#[buildstructor::buildstructor]
-impl SupergraphService {
-    #[builder]
-    pub(crate) fn new(
-        query_planner_service: CachingQueryPlanner<BridgeQueryPlanner>,
-        execution_service_factory: ExecutionServiceFactory,
-        schema: Arc<Schema>,
-        notify: Notify<String, graphql::Response>,
-    ) -> Self {
-        SupergraphService {
-            query_planner_service,
-            execution_service_factory,
-            schema,
-            notify,
-        }
-    }
-}
-
-impl Service<SupergraphRequest> for SupergraphService {
-    type Response = SupergraphResponse;
-    type Error = BoxError;
-    type Future = BoxFuture<'static, Result<Self::Response, Self::Error>>;
-
-    fn poll_ready(&mut self, cx: &mut std::task::Context<'_>) -> Poll<Result<(), Self::Error>> {
-        self.query_planner_service
-            .poll_ready(cx)
-            .map_err(|err| err.into())
-    }
-
-    fn call(&mut self, req: SupergraphRequest) -> Self::Future {
-        // Consume our cloned services and allow ownership to be transferred to the async block.
-        let clone = self.query_planner_service.clone();
-
-        let planning = std::mem::replace(&mut self.query_planner_service, clone);
-
-        let schema = self.schema.clone();
-
-        let context_cloned = req.context.clone();
-        let fut = service_call(
-            planning,
-            self.execution_service_factory.clone(),
-            schema,
-            req,
-            self.notify.clone(),
-        )
-        .or_else(|error: BoxError| async move {
-            let errors = vec![crate::error::Error {
-                message: error.to_string(),
-                extensions: serde_json_bytes::json!({
-                    "code": "INTERNAL_SERVER_ERROR",
-                })
-                .as_object()
-                .unwrap()
-                .to_owned(),
-                ..Default::default()
-            }];
-
-            Ok(SupergraphResponse::builder()
-                .errors(errors)
-                .status_code(StatusCode::INTERNAL_SERVER_ERROR)
-                .context(context_cloned)
-                .build()
-                .expect("building a response like this should not fail"))
-        });
-
-        Box::pin(fut)
-    }
-}
-
-async fn service_call(
-    planning: CachingQueryPlanner<BridgeQueryPlanner>,
-    execution_service_factory: ExecutionServiceFactory,
-    schema: Arc<Schema>,
-    req: SupergraphRequest,
-    notify: Notify<String, graphql::Response>,
-) -> Result<SupergraphResponse, BoxError> {
-    let context = req.context;
-    let body = req.supergraph_request.body();
-    let variables = body.variables.clone();
-
-    let QueryPlannerResponse {
-        content,
-        context,
-        errors,
-    } = match plan_query(
-        planning,
-        body.operation_name.clone(),
-        context.clone(),
-        schema.clone(),
-        req.supergraph_request
-            .body()
-            .query
-            .clone()
-            .expect("query presence was checked before"),
-    )
-    .await
-    {
-        Ok(resp) => resp,
-        Err(err) => match err.into_graphql_errors() {
-            Ok(gql_errors) => {
-                return Ok(SupergraphResponse::builder()
-                    .context(context)
-                    .errors(gql_errors)
-                    .status_code(StatusCode::BAD_REQUEST) // If it's a graphql error we return a status code 400
-                    .build()
-                    .expect("this response build must not fail"));
-            }
-            Err(err) => return Err(err.into()),
-        },
-    };
-
-    if !errors.is_empty() {
-        return Ok(SupergraphResponse::builder()
-            .context(context)
-            .errors(errors)
-            .status_code(StatusCode::BAD_REQUEST) // If it's a graphql error we return a status code 400
-            .build()
-            .expect("this response build must not fail"));
-    }
-
-    match content {
-        Some(QueryPlannerContent::Introspection { response }) => Ok(
-            SupergraphResponse::new_from_graphql_response(*response, context),
-        ),
-        Some(QueryPlannerContent::IntrospectionDisabled) => {
-            let mut response = SupergraphResponse::new_from_graphql_response(
-                graphql::Response::builder()
-                    .errors(vec![crate::error::Error::builder()
-                        .message(String::from("introspection has been disabled"))
-                        .extension_code("INTROSPECTION_DISABLED")
-                        .build()])
-                    .build(),
-                context,
-            );
-            *response.response.status_mut() = StatusCode::BAD_REQUEST;
-            Ok(response)
-        }
-
-        Some(QueryPlannerContent::Plan { plan }) => {
-            let operation_name = body.operation_name.clone();
-            let is_deferred = plan.is_deferred(operation_name.as_deref(), &variables);
-            let is_subscription = plan.is_subscription(operation_name.as_deref());
-
-            if let Some(batching) = context.private_entries.lock().get::<Batching>() {
-                if batching.enabled && (is_deferred || is_subscription) {
-                    let message = if is_deferred {
-                        "BATCHING_DEFER_UNSUPPORTED"
-                    } else {
-                        "BATCHING_SUBSCRIPTION_UNSUPPORTED"
-                    };
-                    let mut response = SupergraphResponse::new_from_graphql_response(
-                            graphql::Response::builder()
-                                .errors(vec![crate::error::Error::builder()
-                                    .message(String::from(
-                                        "Deferred responses and subscriptions aren't supported in batches",
-                                    ))
-                                    .extension_code(message)
-                                    .build()])
-                                .build(),
-                            context.clone(),
-                        );
-                    *response.response.status_mut() = StatusCode::NOT_ACCEPTABLE;
-                    return Ok(response);
-                }
-            }
-
-            let ClientRequestAccepts {
-                multipart_defer: accepts_multipart_defer,
-                multipart_subscription: accepts_multipart_subscription,
-                ..
-            } = context
-                .private_entries
-                .lock()
-                .get()
-                .cloned()
-                .unwrap_or_default();
-            let mut subscription_tx = None;
-            if (is_deferred && !accepts_multipart_defer)
-                || (is_subscription && !accepts_multipart_subscription)
-            {
-                let (error_message, error_code) = if is_deferred {
-                    (String::from("the router received a query with the @defer directive but the client does not accept multipart/mixed HTTP responses. To enable @defer support, add the HTTP header 'Accept: multipart/mixed; deferSpec=20220824'"), "DEFER_BAD_HEADER")
-                } else {
-                    (String::from("the router received a query with a subscription but the client does not accept multipart/mixed HTTP responses. To enable subscription support, add the HTTP header 'Accept: multipart/mixed; boundary=graphql; subscriptionSpec=1.0'"), "SUBSCRIPTION_BAD_HEADER")
-                };
-                let mut response = SupergraphResponse::new_from_graphql_response(
-                    graphql::Response::builder()
-                        .errors(vec![crate::error::Error::builder()
-                            .message(error_message)
-                            .extension_code(error_code)
-                            .build()])
-                        .build(),
-                    context,
-                );
-                *response.response.status_mut() = StatusCode::NOT_ACCEPTABLE;
-                Ok(response)
-            } else if let Some(err) = plan.query.validate_variables(body, &schema).err() {
-                let mut res = SupergraphResponse::new_from_graphql_response(err, context);
-                *res.response.status_mut() = StatusCode::BAD_REQUEST;
-                Ok(res)
-            } else {
-                if is_subscription {
-                    let ctx = context.clone();
-                    let (subs_tx, subs_rx) = mpsc::channel(1);
-                    let query_plan = plan.clone();
-                    let execution_service_factory_cloned = execution_service_factory.clone();
-                    let cloned_supergraph_req =
-                        clone_supergraph_request(&req.supergraph_request, context.clone())?;
-                    // Spawn task for subscription
-                    tokio::spawn(async move {
-                        subscription_task(
-                            execution_service_factory_cloned,
-                            ctx,
-                            query_plan,
-                            subs_rx,
-                            notify,
-                            cloned_supergraph_req,
-                        )
-                        .await;
-                    });
-                    subscription_tx = subs_tx.into();
-                }
-
-                let execution_response = execution_service_factory
-                    .create()
-                    .oneshot(
-                        ExecutionRequest::internal_builder()
-                            .supergraph_request(req.supergraph_request)
-                            .query_plan(plan.clone())
-                            .context(context)
-                            .and_subscription_tx(subscription_tx)
-                            .build()
-                            .await,
-                    )
-                    .await?;
-
-                let ExecutionResponse { response, context } = execution_response;
-
-                let (parts, response_stream) = response.into_parts();
-
-                Ok(SupergraphResponse {
-                    context,
-                    response: http::Response::from_parts(parts, response_stream.boxed()),
-                })
-            }
-        }
-        // This should never happen because if we have an empty query plan we should have error in errors vec
-        None => Err(BoxError::from("cannot compute a query plan")),
-    }
-}
-
-pub struct SubscriptionTaskParams {
-    pub(crate) client_sender: tokio::sync::mpsc::Sender<Response>,
-    pub(crate) subscription_handle: SubscriptionHandle,
-    pub(crate) subscription_config: SubscriptionConfig,
-    pub(crate) stream_rx: ReceiverStream<HandleStream<String, graphql::Response>>,
-    pub(crate) service_name: String,
-}
-
-async fn subscription_task(
-    mut execution_service_factory: ExecutionServiceFactory,
-    context: Context,
-    query_plan: Arc<QueryPlan>,
-    mut rx: mpsc::Receiver<SubscriptionTaskParams>,
-    notify: Notify<String, graphql::Response>,
-    supergraph_req: SupergraphRequest,
-) {
-    let sub_params = match rx.recv().await {
-        Some(sub_params) => sub_params,
-        None => {
-            return;
-        }
-    };
-    let subscription_config = sub_params.subscription_config;
-    let subscription_handle = sub_params.subscription_handle;
-    let service_name = sub_params.service_name;
-    let mut receiver = sub_params.stream_rx;
-    let sender = sub_params.client_sender;
-
-    let graphql_document = &query_plan.query.string;
-    // Get the rest of the query_plan to execute for subscription events
-    let query_plan = match &query_plan.root {
-        crate::query_planner::PlanNode::Subscription { rest, .. } => rest.clone().map(|r| {
-            Arc::new(QueryPlan {
-                usage_reporting: query_plan.usage_reporting.clone(),
-                root: *r,
-                formatted_query_plan: query_plan.formatted_query_plan.clone(),
-                query: query_plan.query.clone(),
-            })
-        }),
-        _ => {
-            let _ = sender
-                .send(
-                    graphql::Response::builder()
-                        .error(
-                            graphql::Error::builder()
-                                .message("cannot execute the subscription event")
-                                .extension_code("SUBSCRIPTION_EXECUTION_ERROR")
-                                .build(),
-                        )
-                        .build(),
-                )
-                .await;
-            return;
-        }
-    };
-
-    let limit_is_set = subscription_config.max_opened_subscriptions.is_some();
-    let mut subscription_handle = subscription_handle.clone();
-    let operation_signature = context
-        .private_entries
-        .lock()
-        .get::<UsageReporting>()
-        .map(|usage_reporting| usage_reporting.stats_report_key.clone())
-        .unwrap_or_default();
-
-    let operation_name = context
-        .get::<_, String>(OPERATION_NAME)
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let display_body = context.contains_key(LOGGING_DISPLAY_BODY);
-
-    let mut receiver = match receiver.next().await {
-        Some(receiver) => receiver,
-        None => {
-            tracing::trace!("receiver channel closed");
-            return;
-        }
-    };
-
-    if limit_is_set {
-        OPENED_SUBSCRIPTIONS.fetch_add(1, Ordering::Relaxed);
-    }
-
-    let mut configuration_updated_rx = notify.subscribe_configuration();
-    let mut schema_updated_rx = notify.subscribe_schema();
-
-    let expires_in = crate::plugins::authentication::jwt_expires_in(&supergraph_req.context);
-
-    let mut timeout = Box::pin(tokio::time::sleep(expires_in));
-
-    loop {
-        tokio::select! {
-            // We prefer to specify the order of checks within the select
-            biased;
-            _ = subscription_handle.closed_signal.recv() => {
-                break;
-            }
-            _ = &mut timeout => {
-                let response = Response::builder()
-                    .subscribed(false)
-                    .error(
-                        crate::error::Error::builder()
-                            .message("subscription closed because the JWT has expired")
-                            .extension_code("SUBSCRIPTION_JWT_EXPIRED")
-                            .build(),
-                    )
-                    .build();
-                let _ = sender.send(response).await;
-                break;
-            },
-            message = receiver.next() => {
-                match message {
-                    Some(mut val) => {
-                        if display_body {
-                            tracing::info!(http.request.body = ?val, apollo.subgraph.name = %service_name, "Subscription event body from subgraph {service_name:?}");
-                        }
-                        val.created_at = Some(Instant::now());
-                        let res = dispatch_event(&supergraph_req, &execution_service_factory, query_plan.as_ref(), context.clone(), val, sender.clone())
-                            .instrument(tracing::info_span!(SUBSCRIPTION_EVENT_SPAN_NAME,
-                                graphql.document = graphql_document,
-                                graphql.operation.name = %operation_name,
-                                otel.kind = "INTERNAL",
-                                apollo_private.operation_signature = %operation_signature,
-                                apollo_private.duration_ns = field::Empty,)
-                            ).await;
-                        if let Err(err) = res {
-                                tracing::error!("cannot send the subscription to the client: {err:?}");
-                            break;
-                        }
-                    }
-                    None => break,
-                }
-            }
-            Some(new_configuration) = configuration_updated_rx.next() => {
-                // If the configuration was dropped in the meantime, we ignore this update and will
-                // pick up the next one.
-                if let Some(conf) = new_configuration.upgrade() {
-                    let plugins = match create_plugins(&conf, &execution_service_factory.schema, None).await {
-                        Ok(plugins) => plugins,
-                        Err(err) => {
-                            tracing::error!("cannot re-create plugins with the new configuration (closing existing subscription): {err:?}");
-                            break;
-                        },
-                    };
-                    let subgraph_services = match create_subgraph_services(&plugins, &execution_service_factory.schema, &conf).await {
-                        Ok(subgraph_services) => subgraph_services,
-                        Err(err) => {
-                            tracing::error!("cannot re-create subgraph service with the new configuration (closing existing subscription): {err:?}");
-                            break;
-                        },
-                    };
-                    let plugins = Arc::new(IndexMap::from_iter(plugins));
-                    execution_service_factory = ExecutionServiceFactory { schema: execution_service_factory.schema.clone(), plugins: plugins.clone(), subgraph_service_factory: Arc::new(SubgraphServiceFactory::new(subgraph_services.into_iter().map(|(k, v)| (k, Arc::new(v) as Arc<dyn MakeSubgraphService>)).collect(), plugins.clone())) };
-                }
-            }
-            Some(new_schema) = schema_updated_rx.next() => {
-                if new_schema.raw_sdl != execution_service_factory.schema.raw_sdl {
-                    let _ = sender
-                        .send(
-                            Response::builder()
-                                .subscribed(false)
-                                .error(graphql::Error::builder().message("subscription has been closed due to a schema reload").extension_code("SUBSCRIPTION_SCHEMA_RELOAD").build())
-                                .build(),
-                        )
-                        .await;
-
-                    break;
-                }
-            }
-        }
-    }
-    drop(sender);
-    tracing::trace!("Leaving the task for subscription");
-    if limit_is_set {
-        OPENED_SUBSCRIPTIONS.fetch_sub(1, Ordering::Relaxed);
-    }
-}
-
-async fn dispatch_event(
-    supergraph_req: &SupergraphRequest,
-    execution_service_factory: &ExecutionServiceFactory,
-    query_plan: Option<&Arc<QueryPlan>>,
-    context: Context,
-    mut val: graphql::Response,
-    sender: mpsc::Sender<Response>,
-) -> Result<(), SendError<Response>> {
-    let start = Instant::now();
-    let span = Span::current();
-    let res = match query_plan {
-        Some(query_plan) => {
-            let cloned_supergraph_req = clone_supergraph_request(
-                &supergraph_req.supergraph_request,
-                supergraph_req.context.clone(),
-            )
-            .expect("it's a clone of the original one; qed");
-            let execution_request = ExecutionRequest::internal_builder()
-                .supergraph_request(cloned_supergraph_req.supergraph_request)
-                .query_plan(query_plan.clone())
-                .context(context)
-                .source_stream_value(val.data.take().unwrap_or_default())
-                .build()
-                .await;
-
-            let execution_service = execution_service_factory.create();
-            let execution_response = execution_service.oneshot(execution_request).await;
-            let next_response = match execution_response {
-                Ok(mut execution_response) => execution_response.next_response().await,
-                Err(err) => {
-                    tracing::error!("cannot execute the subscription event: {err:?}");
-                    let _ = sender
-                        .send(
-                            graphql::Response::builder()
-                                .error(
-                                    graphql::Error::builder()
-                                        .message("cannot execute the subscription event")
-                                        .extension_code("SUBSCRIPTION_EXECUTION_ERROR")
-                                        .build(),
-                                )
-                                .build(),
-                        )
-                        .await;
-                    return Ok(());
-                }
-            };
-
-            if let Some(mut next_response) = next_response {
-                next_response.created_at = val.created_at;
-                next_response.subscribed = val.subscribed;
-                val.errors.append(&mut next_response.errors);
-                next_response.errors = val.errors;
-
-                sender.send(next_response).await
-            } else {
-                Ok(())
-            }
-        }
-        None => sender.send(val).await,
-    };
-    span.record(
-        APOLLO_PRIVATE_DURATION_NS,
-        start.elapsed().as_nanos() as i64,
-    );
-
-    res
-}
-
-async fn plan_query(
-    mut planning: CachingQueryPlanner<BridgeQueryPlanner>,
-    operation_name: Option<String>,
-    context: Context,
-    schema: Arc<Schema>,
-    query_str: String,
-) -> Result<QueryPlannerResponse, CacheResolverError> {
-    // FIXME: we have about 80 tests creating a supergraph service and crafting a supergraph request for it
-    // none of those tests create an executable document to put it in the context, and the document cannot be created
-    // from inside the supergraph request fake builder, because it needs a schema matching the query.
-    // So while we are updating the tests to create a document manually, this here will make sure current
-    // tests will pass
-    {
-        let mut entries = context.private_entries.lock();
-        if !entries.contains_key::<ParsedDocument>() {
-            let doc = Query::parse_document(&query_str, &schema, &Configuration::default());
-            Query::validate_query(&schema, &doc.executable)
-                .map_err(crate::error::QueryPlannerError::from)?;
-            entries.insert::<ParsedDocument>(doc);
-        }
-        drop(entries);
-    }
-
-    planning
-        .call(
-            query_planner::CachingRequest::builder()
-                .query(query_str)
-                .and_operation_name(operation_name)
-                .context(context)
-                .build(),
-        )
-        .instrument(tracing::info_span!(
-            QUERY_PLANNING_SPAN_NAME,
-            "otel.kind" = "INTERNAL"
-        ))
-        .await
-}
-
-fn clone_supergraph_request(
-    req: &http::Request<graphql::Request>,
-    context: Context,
-) -> Result<SupergraphRequest, BoxError> {
-    let mut cloned_supergraph_req = SupergraphRequest::builder()
-        .extensions(req.body().extensions.clone())
-        .and_query(req.body().query.clone())
-        .context(context)
-        .method(req.method().clone())
-        .and_operation_name(req.body().operation_name.clone())
-        .uri(req.uri().clone())
-        .variables(req.body().variables.clone());
-
-    for (header_name, header_value) in req.headers().clone() {
-        if let Some(header_name) = header_name {
-            cloned_supergraph_req = cloned_supergraph_req.header(header_name, header_value);
-        }
-    }
-
-    cloned_supergraph_req.build()
-}
-
-/// Builder which generates a plugin pipeline.
-///
-/// This is at the heart of the delegation of responsibility model for the router. A schema,
-/// collection of plugins, collection of subgraph services are assembled to generate a
-/// [`tower::util::BoxCloneService`] capable of processing a router request
-/// through the entire stack to return a response.
-pub(crate) struct PluggableSupergraphServiceBuilder {
-    plugins: Plugins,
-    subgraph_services: Vec<(String, Arc<dyn MakeSubgraphService>)>,
-    configuration: Option<Arc<Configuration>>,
-    planner: BridgeQueryPlanner,
-}
-
-impl PluggableSupergraphServiceBuilder {
-    pub(crate) fn new(planner: BridgeQueryPlanner) -> Self {
-        Self {
-            plugins: Default::default(),
-            subgraph_services: Default::default(),
-            configuration: None,
-            planner,
-        }
-    }
-
-    pub(crate) fn with_dyn_plugin(
-        mut self,
-        plugin_name: String,
-        plugin: Box<dyn DynPlugin>,
-    ) -> PluggableSupergraphServiceBuilder {
-        self.plugins.insert(plugin_name, plugin);
-        self
-    }
-
-    pub(crate) fn with_subgraph_service<S>(
-        mut self,
-        name: &str,
-        service_maker: S,
-    ) -> PluggableSupergraphServiceBuilder
-    where
-        S: MakeSubgraphService,
-    {
-        self.subgraph_services
-            .push((name.to_string(), Arc::new(service_maker)));
-        self
-    }
-
-    pub(crate) fn with_configuration(
-        mut self,
-        configuration: Arc<Configuration>,
-    ) -> PluggableSupergraphServiceBuilder {
-        self.configuration = Some(configuration);
-        self
-    }
-
-    pub(crate) async fn build(self) -> Result<SupergraphCreator, crate::error::ServiceBuildError> {
-        let configuration = self.configuration.unwrap_or_default();
-
-        let schema = self.planner.schema();
-        let query_planner_service = CachingQueryPlanner::new(
-            self.planner,
-            schema.clone(),
-            &configuration,
-            IndexMap::new(),
-        )
-        .await;
-
-        let mut plugins = self.plugins;
-        // Activate the telemetry plugin.
-        // We must NOT fail to go live with the new router from this point as the telemetry plugin activate interacts with globals.
-        for (_, plugin) in plugins.iter_mut() {
-            if let Some(telemetry) = plugin.as_any_mut().downcast_mut::<Telemetry>() {
-                telemetry.activate();
-            }
-        }
-
-        let plugins = Arc::new(plugins);
-
-        let subgraph_service_factory = Arc::new(SubgraphServiceFactory::new(
-            self.subgraph_services,
-            plugins.clone(),
-        ));
-
-        Ok(SupergraphCreator {
-            query_planner_service,
-            subgraph_service_factory,
-            schema,
-            plugins,
-            config: configuration,
-        })
-    }
-}
-
-/// A collection of services and data which may be used to create a "router".
-#[derive(Clone)]
-pub(crate) struct SupergraphCreator {
-    query_planner_service: CachingQueryPlanner<BridgeQueryPlanner>,
-    subgraph_service_factory: Arc<SubgraphServiceFactory>,
-    schema: Arc<Schema>,
-    config: Arc<Configuration>,
-    plugins: Arc<Plugins>,
-}
-
-pub(crate) trait HasPlugins {
-    fn plugins(&self) -> Arc<Plugins>;
-}
-
-impl HasPlugins for SupergraphCreator {
-    fn plugins(&self) -> Arc<Plugins> {
-        self.plugins.clone()
-    }
-}
-
-pub(crate) trait HasSchema {
-    fn schema(&self) -> Arc<Schema>;
-}
-
-impl HasSchema for SupergraphCreator {
-    fn schema(&self) -> Arc<Schema> {
-        Arc::clone(&self.schema)
-    }
-}
-
-pub(crate) trait HasConfig {
-    fn config(&self) -> Arc<Configuration>;
-}
-
-impl HasConfig for SupergraphCreator {
-    fn config(&self) -> Arc<Configuration> {
-        Arc::clone(&self.config)
-    }
-}
-
-impl ServiceFactory<supergraph::Request> for SupergraphCreator {
-    type Service = supergraph::BoxService;
-    fn create(&self) -> Self::Service {
-        self.make().boxed()
-    }
-}
-
-impl SupergraphCreator {
-    pub(crate) fn make(
-        &self,
-    ) -> impl Service<
-        supergraph::Request,
-        Response = supergraph::Response,
-        Error = BoxError,
-        Future = BoxFuture<'static, supergraph::ServiceResult>,
-    > + Send {
-        let supergraph_service = SupergraphService::builder()
-            .query_planner_service(self.query_planner_service.clone())
-            .execution_service_factory(ExecutionServiceFactory {
-                schema: self.schema.clone(),
-                plugins: self.plugins.clone(),
-                subgraph_service_factory: self.subgraph_service_factory.clone(),
-            })
-            .schema(self.schema.clone())
-            .notify(self.config.notify.clone())
-            .build();
-
-        let shaping = self
-            .plugins
-            .iter()
-            .find(|i| i.0.as_str() == APOLLO_TRAFFIC_SHAPING)
-            .and_then(|plugin| plugin.1.as_any().downcast_ref::<TrafficShaping>())
-            .expect("traffic shaping should always be part of the plugin list");
-
-        let supergraph_service = AllowOnlyHttpPostMutationsLayer::default()
-            .layer(shaping.supergraph_service_internal(supergraph_service));
-
-        ServiceBuilder::new()
-            .layer(content_negotiation::SupergraphLayer::default())
-            .service(
-                self.plugins
-                    .iter()
-                    .rev()
-                    .fold(supergraph_service.boxed(), |acc, (_, e)| {
-                        e.supergraph_service(acc)
-                    }),
-            )
-    }
-
-    pub(crate) async fn cache_keys(&self, count: Option<usize>) -> Vec<WarmUpCachingQueryKey> {
-        self.query_planner_service.cache_keys(count).await
-    }
-
-    pub(crate) fn planner(&self) -> Arc<Planner<QueryPlanResult>> {
-        self.query_planner_service.planner()
-    }
-
-    pub(crate) async fn warm_up_query_planner(
-        &mut self,
-        query_parser: &QueryAnalysisLayer,
-        persisted_query_layer: &PersistedQueryLayer,
-        cache_keys: Vec<WarmUpCachingQueryKey>,
-    ) {
-        self.query_planner_service
-            .warm_up(query_parser, persisted_query_layer, cache_keys)
-            .await
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-    use std::time::Duration;
-
-    use http::HeaderValue;
-    use tower::ServiceExt;
-
-    use super::*;
-    use crate::plugin::test::MockSubgraph;
-    use crate::services::subgraph;
-    use crate::services::supergraph;
-    use crate::test_harness::MockedSubgraphs;
-    use crate::Notify;
-    use crate::TestHarness;
-
-    const SCHEMA: &str = r#"schema
+const SCHEMA: &str = r#"schema
         @core(feature: "https://specs.apollo.dev/core/v0.1")
         @core(feature: "https://specs.apollo.dev/join/v0.1")
         @core(feature: "https://specs.apollo.dev/inaccessible/v0.1")
@@ -906,9 +64,9 @@ mod tests {
        suborga: [Organization]
    }"#;
 
-    #[tokio::test]
-    async fn nullability_formatting() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn nullability_formatting() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": null }}}}
@@ -916,35 +74,35 @@ mod tests {
         ("orga", MockSubgraph::default())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query("query { currentUser { activeOrganization { id creatorUser { name } } } }")
-            .context(defer_context())
-            // Request building here
-            .build()
-            .unwrap();
-        let response = service
-            .oneshot(request)
-            .await
-            .unwrap()
-            .next_response()
-            .await
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .query("query { currentUser { activeOrganization { id creatorUser { name } } } }")
+        .context(defer_context())
+        // Request building here
+        .build()
+        .unwrap();
+    let response = service
+        .oneshot(request)
+        .await
+        .unwrap()
+        .next_response()
+        .await
+        .unwrap();
 
-        insta::assert_json_snapshot!(response);
-    }
+    insta::assert_json_snapshot!(response);
+}
 
-    #[tokio::test]
-    async fn nullability_bubbling() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn nullability_bubbling() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": {} }}}}
@@ -952,36 +110,34 @@ mod tests {
         ("orga", MockSubgraph::default())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                "query { currentUser { activeOrganization { nonNullId creatorUser { name } } } }",
-            )
-            .build()
-            .unwrap();
-        let response = service
-            .oneshot(request)
-            .await
-            .unwrap()
-            .next_response()
-            .await
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query("query { currentUser { activeOrganization { nonNullId creatorUser { name } } } }")
+        .build()
+        .unwrap();
+    let response = service
+        .oneshot(request)
+        .await
+        .unwrap()
+        .next_response()
+        .await
+        .unwrap();
 
-        insta::assert_json_snapshot!(response);
-    }
+    insta::assert_json_snapshot!(response);
+}
 
-    #[tokio::test]
-    async fn errors_on_deferred_responses() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn errors_on_deferred_responses() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{__typename id}}"}},
                 serde_json::json!{{"data": {"currentUser": { "__typename": "User", "id": "0" }}}}
@@ -1010,31 +166,31 @@ mod tests {
         ("orga", MockSubgraph::default())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query("query { currentUser { id  ...@defer { name } } }")
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query("query { currentUser { id  ...@defer { name } } }")
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn errors_from_primary_on_deferred_responses() {
-        let schema = r#"
+#[tokio::test]
+async fn errors_from_primary_on_deferred_responses() {
+    let schema = r#"
         schema
           @link(url: "https://specs.apollo.dev/link/v1.0")
           @link(url: "https://specs.apollo.dev/join/v0.2", for: EXECUTION)
@@ -1075,7 +231,7 @@ mod tests {
           computer(id: ID!): Computer
         }"#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
         ("computers", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{__typename id}}"}},
                 serde_json::json!{{"data": {"currentUser": { "__typename": "User", "id": "0" }}}}
@@ -1106,19 +262,19 @@ mod tests {
             ).build()),
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                r#"query {
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(
+            r#"query {
                 computer(id: "Computer1") {
                   id
                   ...ComputerErrorField @defer
@@ -1127,20 +283,20 @@ mod tests {
               fragment ComputerErrorField on Computer {
                 errorField
               }"#,
-            )
-            .build()
-            .unwrap();
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn deferred_fragment_bounds_nullability() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn deferred_fragment_bounds_nullability() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1196,16 +352,16 @@ mod tests {
         ).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
         .context(defer_context())
         .query(
                 "query { currentUser { activeOrganization { id  suborga { id ...@defer { nonNullId } } } } }",
@@ -1213,16 +369,16 @@ mod tests {
             .build()
             .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn errors_on_incremental_responses() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn errors_on_incremental_responses() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1278,16 +434,16 @@ mod tests {
         ).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
         .context(defer_context())
         .query(
                 "query { currentUser { activeOrganization { id  suborga { id ...@defer { name } } } } }",
@@ -1295,16 +451,16 @@ mod tests {
             .build()
             .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn root_typename_with_defer() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn root_typename_with_defer() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1350,16 +506,16 @@ mod tests {
         ).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .context(defer_context())
             .query(
                 "query { __typename currentUser { activeOrganization { id  suborga { id ...@defer { name } } } } }",
@@ -1367,25 +523,25 @@ mod tests {
             .build()
             .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        assert_eq!(
-            res.data.as_ref().unwrap().get("__typename"),
-            Some(&serde_json_bytes::Value::String("Query".into()))
-        );
-        insta::assert_json_snapshot!(res);
+    let mut stream = service.oneshot(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    assert_eq!(
+        res.data.as_ref().unwrap().get("__typename"),
+        Some(&serde_json_bytes::Value::String("Query".into()))
+    );
+    insta::assert_json_snapshot!(res);
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn subscription_with_callback() {
-        let mut notify = Notify::builder().build();
-        let (handle, _) = notify
-            .create_or_subscribe("TEST_TOPIC".to_string(), false)
-            .await
-            .unwrap();
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn subscription_with_callback() {
+    let mut notify = Notify::builder().build();
+    let (handle, _) = notify
+        .create_or_subscribe("TEST_TOPIC".to_string(), false)
+        .await
+        .unwrap();
+    let subgraphs = MockedSubgraphs([
             ("user", MockSubgraph::builder().with_json(
                     serde_json::json!{{"query":"subscription{userWasCreated{name activeOrganization{__typename id}}}"}},
                     serde_json::json!{{"data": {"userWasCreated": { "__typename": "User", "id": "1", "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1409,53 +565,53 @@ mod tests {
             ).build())
         ].into_iter().collect());
 
-        let mut configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "subscription": { "enabled": true, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
-        configuration.notify = notify.clone();
-        let service = TestHarness::builder()
-            .configuration(Arc::new(configuration))
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let mut configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "subscription": { "enabled": true, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
+    configuration.notify = notify.clone();
+    let service = TestHarness::builder()
+        .configuration(Arc::new(configuration))
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .query(
                 "subscription { userWasCreated { name activeOrganization { id  suborga { id name } } } }",
             )
             .context(subscription_context())
             .build()
             .unwrap();
-        let mut stream = service.oneshot(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        insta::assert_json_snapshot!(res);
-        notify.broadcast(graphql::Response::builder().data(serde_json_bytes::json!({"userWasCreated": { "name": "test", "activeOrganization": { "__typename": "Organization", "id": "0" }}})).build()).await.unwrap();
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-        // error happened
-        notify
-            .broadcast(
-                graphql::Response::builder()
-                    .error(
-                        graphql::Error::builder()
-                            .message("cannot fetch the name")
-                            .extension_code("INVALID")
-                            .build(),
-                    )
-                    .build(),
-            )
-            .await
-            .unwrap();
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    let mut stream = service.oneshot(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    insta::assert_json_snapshot!(res);
+    notify.broadcast(graphql::Response::builder().data(serde_json_bytes::json!({"userWasCreated": { "name": "test", "activeOrganization": { "__typename": "Organization", "id": "0" }}})).build()).await.unwrap();
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    // error happened
+    notify
+        .broadcast(
+            graphql::Response::builder()
+                .error(
+                    graphql::Error::builder()
+                        .message("cannot fetch the name")
+                        .extension_code("INVALID")
+                        .build(),
+                )
+                .build(),
+        )
+        .await
+        .unwrap();
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn subscription_callback_schema_reload() {
-        let mut notify = Notify::builder().build();
-        let (handle, _) = notify
-            .create_or_subscribe("TEST_TOPIC".to_string(), false)
-            .await
-            .unwrap();
-        let orga_subgraph = MockSubgraph::builder().with_json(
+#[tokio::test]
+async fn subscription_callback_schema_reload() {
+    let mut notify = Notify::builder().build();
+    let (handle, _) = notify
+        .create_or_subscribe("TEST_TOPIC".to_string(), false)
+        .await
+        .unwrap();
+    let orga_subgraph = MockSubgraph::builder().with_json(
                 serde_json::json!{{
                     "query":"query($representations:[_Any!]!){_entities(representations:$representations){...on Organization{suborga{id name}}}}",
                     "variables": {
@@ -1476,7 +632,7 @@ mod tests {
                 assert_eq!(req.subgraph_request.headers().get("x-test").unwrap(), HeaderValue::from_static("test"));
                 req
             });
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             ("user", MockSubgraph::builder().with_json(
                     serde_json::json!{{"query":"subscription{userWasCreated{name activeOrganization{__typename id}}}"}},
                     serde_json::json!{{"data": {"userWasCreated": { "__typename": "User", "id": "1", "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1484,18 +640,18 @@ mod tests {
             ("orga", orga_subgraph)
         ].into_iter().collect());
 
-        let mut configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "headers": {"all": {"request": [{"propagate": {"named": "x-test"}}]}}, "subscription": { "enabled": true, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
-        configuration.notify = notify.clone();
-        let configuration = Arc::new(configuration);
-        let service = TestHarness::builder()
-            .configuration(configuration.clone())
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let mut configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "headers": {"all": {"request": [{"propagate": {"named": "x-test"}}]}}, "subscription": { "enabled": true, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
+    configuration.notify = notify.clone();
+    let configuration = Arc::new(configuration);
+    let service = TestHarness::builder()
+        .configuration(configuration.clone())
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .query(
                 "subscription { userWasCreated { name activeOrganization { id  suborga { id name } } } }",
             )
@@ -1503,33 +659,33 @@ mod tests {
             .context(subscription_context())
             .build()
             .unwrap();
-        let mut stream = service.oneshot(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        insta::assert_json_snapshot!(res);
-        notify.broadcast(graphql::Response::builder().data(serde_json_bytes::json!({"userWasCreated": { "name": "test", "activeOrganization": { "__typename": "Organization", "id": "0" }}})).build()).await.unwrap();
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    let mut stream = service.oneshot(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    insta::assert_json_snapshot!(res);
+    notify.broadcast(graphql::Response::builder().data(serde_json_bytes::json!({"userWasCreated": { "name": "test", "activeOrganization": { "__typename": "Organization", "id": "0" }}})).build()).await.unwrap();
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
 
-        let new_schema = format!("{SCHEMA}  ");
-        // reload schema
-        let schema = Schema::parse(&new_schema, &configuration).unwrap();
-        notify.broadcast_schema(Arc::new(schema));
-        insta::assert_json_snapshot!(tokio::time::timeout(
-            Duration::from_secs(1),
-            stream.next_response()
-        )
+    let new_schema = format!("{SCHEMA}  ");
+    // reload schema
+    let schema = Schema::parse(&new_schema, &configuration).unwrap();
+    notify.broadcast_schema(Arc::new(schema));
+    insta::assert_json_snapshot!(tokio::time::timeout(
+        Duration::from_secs(1),
+        stream.next_response()
+    )
+    .await
+    .unwrap()
+    .unwrap());
+}
+
+#[tokio::test]
+async fn subscription_with_callback_with_limit() {
+    let mut notify = Notify::builder().build();
+    let (handle, _) = notify
+        .create_or_subscribe("TEST_TOPIC".to_string(), false)
         .await
-        .unwrap()
-        .unwrap());
-    }
-
-    #[tokio::test]
-    async fn subscription_with_callback_with_limit() {
-        let mut notify = Notify::builder().build();
-        let (handle, _) = notify
-            .create_or_subscribe("TEST_TOPIC".to_string(), false)
-            .await
-            .unwrap();
-        let subgraphs = MockedSubgraphs([
+        .unwrap();
+    let subgraphs = MockedSubgraphs([
             ("user", MockSubgraph::builder().with_json(
                     serde_json::json!{{"query":"subscription{userWasCreated{name activeOrganization{__typename id}}}"}},
                     serde_json::json!{{"data": {"userWasCreated": { "__typename": "User", "id": "1", "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1553,97 +709,97 @@ mod tests {
             ).build())
         ].into_iter().collect());
 
-        let mut configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "subscription": { "enabled": true, "max_opened_subscriptions": 1, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
-        configuration.notify = notify.clone();
-        let mut service = TestHarness::builder()
-            .configuration(Arc::new(configuration))
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let mut configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "subscription": { "enabled": true, "max_opened_subscriptions": 1, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
+    configuration.notify = notify.clone();
+    let mut service = TestHarness::builder()
+        .configuration(Arc::new(configuration))
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .query(
                 "subscription { userWasCreated { name activeOrganization { id  suborga { id name } } } }",
             )
             .context(subscription_context())
             .build()
             .unwrap();
-        let mut stream = service.ready().await.unwrap().call(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        insta::assert_json_snapshot!(res);
-        notify.broadcast(graphql::Response::builder().data(serde_json_bytes::json!({"userWasCreated": { "name": "test", "activeOrganization": { "__typename": "Organization", "id": "0" }}})).build()).await.unwrap();
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-        // error happened
-        notify
-            .broadcast(
-                graphql::Response::builder()
-                    .error(
-                        graphql::Error::builder()
-                            .message("cannot fetch the name")
-                            .extension_code("INVALID")
-                            .build(),
-                    )
-                    .build(),
-            )
-            .await
-            .unwrap();
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-        let request = supergraph::Request::fake_builder()
+    let mut stream = service.ready().await.unwrap().call(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    insta::assert_json_snapshot!(res);
+    notify.broadcast(graphql::Response::builder().data(serde_json_bytes::json!({"userWasCreated": { "name": "test", "activeOrganization": { "__typename": "Organization", "id": "0" }}})).build()).await.unwrap();
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    // error happened
+    notify
+        .broadcast(
+            graphql::Response::builder()
+                .error(
+                    graphql::Error::builder()
+                        .message("cannot fetch the name")
+                        .extension_code("INVALID")
+                        .build(),
+                )
+                .build(),
+        )
+        .await
+        .unwrap();
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    let request = supergraph::Request::fake_builder()
             .query(
                 "subscription { userWasCreated { name activeOrganization { id  suborga { id name } } } }",
             )
             .context(subscription_context())
             .build()
             .unwrap();
-        let mut stream_2 = service.ready().await.unwrap().call(request).await.unwrap();
-        let res = stream_2.next_response().await.unwrap();
-        assert!(!res.errors.is_empty());
-        insta::assert_json_snapshot!(res);
-        drop(stream);
-        drop(stream_2);
-        let request = supergraph::Request::fake_builder()
+    let mut stream_2 = service.ready().await.unwrap().call(request).await.unwrap();
+    let res = stream_2.next_response().await.unwrap();
+    assert!(!res.errors.is_empty());
+    insta::assert_json_snapshot!(res);
+    drop(stream);
+    drop(stream_2);
+    let request = supergraph::Request::fake_builder()
             .query(
                 "subscription { userWasCreated { name activeOrganization { id  suborga { id name } } } }",
             )
             .context(subscription_context())
             .build()
             .unwrap();
-        // Wait a bit to ensure all the closed signals has been triggered
-        tokio::time::sleep(Duration::from_millis(100)).await;
-        let mut stream_2 = service.ready().await.unwrap().call(request).await.unwrap();
-        let res = stream_2.next_response().await.unwrap();
-        assert!(res.errors.is_empty());
-    }
+    // Wait a bit to ensure all the closed signals has been triggered
+    tokio::time::sleep(Duration::from_millis(100)).await;
+    let mut stream_2 = service.ready().await.unwrap().call(request).await.unwrap();
+    let res = stream_2.next_response().await.unwrap();
+    assert!(res.errors.is_empty());
+}
 
-    #[tokio::test]
-    async fn subscription_without_header() {
-        let subgraphs = MockedSubgraphs(HashMap::new());
-        let configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "subscription": { "enabled": true, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
-        let service = TestHarness::builder()
-            .configuration(Arc::new(configuration))
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+#[tokio::test]
+async fn subscription_without_header() {
+    let subgraphs = MockedSubgraphs(HashMap::new());
+    let configuration: Configuration = serde_json::from_value(serde_json::json!({"include_subgraph_errors": { "all": true }, "subscription": { "enabled": true, "mode": {"preview_callback": {"public_url": "http://localhost:4545"}}}})).unwrap();
+    let service = TestHarness::builder()
+        .configuration(Arc::new(configuration))
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .query(
                 "subscription { userWasCreated { name activeOrganization { id  suborga { id name } } } }",
             )
             .build()
             .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        insta::assert_json_snapshot!(res);
-    }
+    let mut stream = service.oneshot(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    insta::assert_json_snapshot!(res);
+}
 
-    #[tokio::test]
-    async fn root_typename_with_defer_and_empty_first_response() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn root_typename_with_defer_and_empty_first_response() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1689,16 +845,16 @@ mod tests {
         ).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .context(defer_context())
             .query(
                 "query { __typename ... @defer { currentUser { activeOrganization { id  suborga { id name } } } } }",
@@ -1706,20 +862,20 @@ mod tests {
             .build()
             .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        assert_eq!(
-            res.data.as_ref().unwrap().get("__typename"),
-            Some(&serde_json_bytes::Value::String("Query".into()))
-        );
+    let mut stream = service.oneshot(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    assert_eq!(
+        res.data.as_ref().unwrap().get("__typename"),
+        Some(&serde_json_bytes::Value::String("Query".into()))
+    );
 
-        // Must have 2 chunks
-        let _ = stream.next_response().await.unwrap();
-    }
+    // Must have 2 chunks
+    let _ = stream.next_response().await.unwrap();
+}
 
-    #[tokio::test]
-    async fn root_typename_with_defer_in_defer() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn root_typename_with_defer_in_defer() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "activeOrganization": { "__typename": "Organization", "id": "0" } }}}}
@@ -1743,16 +899,16 @@ mod tests {
         ).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
+    let request = supergraph::Request::fake_builder()
             .context(defer_context())
             .query(
                 "query { ...@defer { __typename currentUser { activeOrganization { id  suborga { id name } } } } }",
@@ -1760,25 +916,25 @@ mod tests {
             .build()
             .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let res = stream.next_response().await.unwrap();
-        assert_eq!(res.errors, []);
-        let res = stream.next_response().await.unwrap();
-        assert_eq!(
-            res.incremental
-                .get(0)
-                .unwrap()
-                .data
-                .as_ref()
-                .unwrap()
-                .get("__typename"),
-            Some(&serde_json_bytes::Value::String("Query".into()))
-        );
-    }
+    let mut stream = service.oneshot(request).await.unwrap();
+    let res = stream.next_response().await.unwrap();
+    assert_eq!(res.errors, []);
+    let res = stream.next_response().await.unwrap();
+    assert_eq!(
+        res.incremental
+            .get(0)
+            .unwrap()
+            .data
+            .as_ref()
+            .unwrap()
+            .get("__typename"),
+        Some(&serde_json_bytes::Value::String("Query".into()))
+    );
+}
 
-    #[tokio::test]
-    async fn query_reconstruction() {
-        let schema = r#"schema
+#[tokio::test]
+async fn query_reconstruction() {
+    let schema = r#"schema
     @link(url: "https://specs.apollo.dev/link/v1.0")
     @link(url: "https://specs.apollo.dev/join/v0.2", for: EXECUTION)
     @link(url: "https://specs.apollo.dev/tag/v0.2")
@@ -1844,20 +1000,20 @@ mod tests {
   }
   "#;
 
-        // this test does not need to generate a valid response, it is only here to check
-        // that the router does not panic when reconstructing the query for the deferred part
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .build_supergraph()
-            .await
-            .unwrap();
+    // this test does not need to generate a valid response, it is only here to check
+    // that the router does not panic when reconstructing the query for the deferred part
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                r#"mutation ($userId: ID!) {
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(
+            r#"mutation ($userId: ID!) {
                     makePayment(userId: $userId) {
                       id
                       ... @defer {
@@ -1867,20 +1023,20 @@ mod tests {
                       }
                     }
                   }"#,
-            )
-            .build()
-            .unwrap();
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    // if a deferred response falls under a path that was nullified in the primary response,
-    // the deferred response must not be sent
-    #[tokio::test]
-    async fn filter_nullified_deferred_responses() {
-        let subgraphs = MockedSubgraphs([
+// if a deferred response falls under a path that was nullified in the primary response,
+// the deferred response must not be sent
+#[tokio::test]
+async fn filter_nullified_deferred_responses() {
+    let subgraphs = MockedSubgraphs([
         ("user", MockSubgraph::builder()
         .with_json(
             serde_json::json!{{"query":"{currentUser{__typename name id}}"}},
@@ -1958,18 +1114,18 @@ mod tests {
                         }}).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query(
-                r#"query {
+    let request = supergraph::Request::fake_builder()
+        .query(
+            r#"query {
                 currentUser {
                     name
                     ... @defer {
@@ -1985,27 +1141,27 @@ mod tests {
                     }
                 }
             }"#,
-            )
-            .context(defer_context())
-            .build()
-            .unwrap();
-        let mut response = service.oneshot(request).await.unwrap();
+        )
+        .context(defer_context())
+        .build()
+        .unwrap();
+    let mut response = service.oneshot(request).await.unwrap();
 
-        let primary = response.next_response().await.unwrap();
-        insta::assert_json_snapshot!(primary);
+    let primary = response.next_response().await.unwrap();
+    insta::assert_json_snapshot!(primary);
 
-        let deferred = response.next_response().await.unwrap();
-        insta::assert_json_snapshot!(deferred);
+    let deferred = response.next_response().await.unwrap();
+    insta::assert_json_snapshot!(deferred);
 
-        // the last deferred response was replace with an empty response,
-        // to still have one containing has_next = false
-        let last = response.next_response().await.unwrap();
-        insta::assert_json_snapshot!(last);
-    }
+    // the last deferred response was replace with an empty response,
+    // to still have one containing has_next = false
+    let last = response.next_response().await.unwrap();
+    insta::assert_json_snapshot!(last);
+}
 
-    #[tokio::test]
-    async fn reconstruct_deferred_query_under_interface() {
-        let schema = r#"schema
+#[tokio::test]
+async fn reconstruct_deferred_query_under_interface() {
+    let schema = r#"schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
             @link(url: "https://specs.apollo.dev/join/v0.2", for: EXECUTION)
             @link(url: "https://specs.apollo.dev/tag/v0.2")
@@ -2074,7 +1230,7 @@ mod tests {
                 name: String!  @join__field(graph: USER)
             }"#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             ("user", MockSubgraph::builder().with_json(
             serde_json::json!{{"query":"{me{__typename ...on User{id fullName memberships{permission account{__typename id}}}}}"}},
             serde_json::json!{{"data": {"me": {
@@ -2109,19 +1265,19 @@ mod tests {
             }}).build()),
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                r#"query {
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(
+            r#"query {
                     me {
                       ... on User {
                         id
@@ -2137,39 +1293,39 @@ mod tests {
                       }
                     }
                   }"#,
-            )
-            .build()
-            .unwrap();
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    fn subscription_context() -> Context {
-        let context = Context::new();
-        context.private_entries.lock().insert(ClientRequestAccepts {
-            multipart_subscription: true,
-            ..Default::default()
-        });
+fn subscription_context() -> Context {
+    let context = Context::new();
+    context.private_entries.lock().insert(ClientRequestAccepts {
+        multipart_subscription: true,
+        ..Default::default()
+    });
 
-        context
-    }
+    context
+}
 
-    fn defer_context() -> Context {
-        let context = Context::new();
-        context.private_entries.lock().insert(ClientRequestAccepts {
-            multipart_defer: true,
-            ..Default::default()
-        });
+fn defer_context() -> Context {
+    let context = Context::new();
+    context.private_entries.lock().insert(ClientRequestAccepts {
+        multipart_defer: true,
+        ..Default::default()
+    });
 
-        context
-    }
+    context
+}
 
-    #[tokio::test]
-    async fn interface_object_typename_rewrites() {
-        let schema = r#"
+#[tokio::test]
+async fn interface_object_typename_rewrites() {
+    let schema = r#"
             schema
               @link(url: "https://specs.apollo.dev/link/v1.0")
               @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
@@ -2237,7 +1393,7 @@ mod tests {
             }
         "#;
 
-        let query = r#"
+    let query = r#"
           {
             iFromS1 {
               ... on A {
@@ -2247,7 +1403,7 @@ mod tests {
           }
         "#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             ("S1", MockSubgraph::builder()
                 .with_json(
                     serde_json::json! {{
@@ -2274,32 +1430,32 @@ mod tests {
                 .build()),
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query(query)
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .query(query)
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let response = stream.next_response().await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
+    let response = stream.next_response().await.unwrap();
 
-        assert_eq!(
-            serde_json::to_value(&response.data).unwrap(),
-            serde_json::json!({ "iFromS1": { "y": 42 } }),
-        );
-    }
+    assert_eq!(
+        serde_json::to_value(&response.data).unwrap(),
+        serde_json::json!({ "iFromS1": { "y": 42 } }),
+    );
+}
 
-    #[tokio::test]
-    async fn interface_object_response_processing() {
-        let schema = r#"
+#[tokio::test]
+async fn interface_object_response_processing() {
+    let schema = r#"
           schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
             @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
@@ -2379,7 +1535,7 @@ mod tests {
           }
         "#;
 
-        let query = r#"
+    let query = r#"
           {
             allReviewedProducts {
               id
@@ -2388,7 +1544,7 @@ mod tests {
           }
         "#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             ("products", MockSubgraph::builder()
                 .with_json(
                     serde_json::json! {{
@@ -2412,45 +1568,45 @@ mod tests {
                 .build()),
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query(query)
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .query(query)
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let response = stream.next_response().await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
+    let response = stream.next_response().await.unwrap();
 
-        assert_eq!(
-            serde_json::to_value(&response.data).unwrap(),
-            serde_json::json!({ "allReviewedProducts": [ {"id": "1", "price": 12.99}, {"id": "2", "price": 14.99} ]}),
-        );
-    }
+    assert_eq!(
+        serde_json::to_value(&response.data).unwrap(),
+        serde_json::json!({ "allReviewedProducts": [ {"id": "1", "price": 12.99}, {"id": "2", "price": 14.99} ]}),
+    );
+}
 
-    #[tokio::test]
-    async fn only_query_interface_object_subgraph() {
-        // This test has 2 subgraphs, one with an interface and another with that interface
-        // declared as an @interfaceObject. It then sends a query that can be entirely
-        // fulfilled by the @interfaceObject subgraph (in particular, it doesn't request
-        // __typename; if it did, it would force a query on the other subgraph to obtain
-        // the actual implementation type).
-        // The specificity here is that the final in-memory result will not have a __typename
-        // _despite_ being the parent type of that result being an interface. Which is fine
-        // since __typename is not requested, and so there is no need to known the actual
-        // __typename, but this is something that never happen outside of @interfaceObject
-        // (usually, results whose parent type is an abstract type (say an interface) are always
-        // queried internally with their __typename). And so this test make sure that the
-        // post-processing done by the router on the result handle this correctly.
+#[tokio::test]
+async fn only_query_interface_object_subgraph() {
+    // This test has 2 subgraphs, one with an interface and another with that interface
+    // declared as an @interfaceObject. It then sends a query that can be entirely
+    // fulfilled by the @interfaceObject subgraph (in particular, it doesn't request
+    // __typename; if it did, it would force a query on the other subgraph to obtain
+    // the actual implementation type).
+    // The specificity here is that the final in-memory result will not have a __typename
+    // _despite_ being the parent type of that result being an interface. Which is fine
+    // since __typename is not requested, and so there is no need to known the actual
+    // __typename, but this is something that never happen outside of @interfaceObject
+    // (usually, results whose parent type is an abstract type (say an interface) are always
+    // queried internally with their __typename). And so this test make sure that the
+    // post-processing done by the router on the result handle this correctly.
 
-        let schema = r#"
+    let schema = r#"
           schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
             @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
@@ -2524,7 +1680,7 @@ mod tests {
           }
         "#;
 
-        let query = r#"
+    let query = r#"
           {
             iFromS2 {
               y
@@ -2532,58 +1688,58 @@ mod tests {
           }
         "#;
 
-        let subgraphs = MockedSubgraphs(
-            [
-                (
-                    "S1",
-                    MockSubgraph::builder()
-                        // This test makes no queries to S1, only to S2
-                        .build(),
-                ),
-                (
-                    "S2",
-                    MockSubgraph::builder()
-                        .with_json(
-                            serde_json::json! {{
-                                "query": "{iFromS2{y}}",
-                            }},
-                            serde_json::json! {{
-                                "data": {"iFromS2":{"y":20}}
-                            }},
-                        )
-                        .build(),
-                ),
-            ]
-            .into_iter()
-            .collect(),
-        );
+    let subgraphs = MockedSubgraphs(
+        [
+            (
+                "S1",
+                MockSubgraph::builder()
+                    // This test makes no queries to S1, only to S2
+                    .build(),
+            ),
+            (
+                "S2",
+                MockSubgraph::builder()
+                    .with_json(
+                        serde_json::json! {{
+                            "query": "{iFromS2{y}}",
+                        }},
+                        serde_json::json! {{
+                            "data": {"iFromS2":{"y":20}}
+                        }},
+                    )
+                    .build(),
+            ),
+        ]
+        .into_iter()
+        .collect(),
+    );
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query(query)
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .query(query)
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let response = stream.next_response().await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
+    let response = stream.next_response().await.unwrap();
 
-        assert_eq!(
-            serde_json::to_value(&response.data).unwrap(),
-            serde_json::json!({ "iFromS2": { "y": 20 } }),
-        );
-    }
+    assert_eq!(
+        serde_json::to_value(&response.data).unwrap(),
+        serde_json::json!({ "iFromS2": { "y": 20 } }),
+    );
+}
 
-    #[tokio::test]
-    async fn aliased_subgraph_data_rewrites_on_root_fetch() {
-        let schema = r#"
+#[tokio::test]
+async fn aliased_subgraph_data_rewrites_on_root_fetch() {
+    let schema = r#"
           schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
             @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
@@ -2645,7 +1801,7 @@ mod tests {
           }
         "#;
 
-        let query = r#"
+    let query = r#"
           {
             us {
               f
@@ -2653,7 +1809,7 @@ mod tests {
           }
         "#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             ("S1", MockSubgraph::builder()
                 .with_json(
                     serde_json::json! {{
@@ -2681,32 +1837,32 @@ mod tests {
                 .build()),
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query(query)
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .query(query)
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let response = stream.next_response().await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
+    let response = stream.next_response().await.unwrap();
 
-        assert_eq!(
-            serde_json::to_value(&response.data).unwrap(),
-            serde_json::json!({"us": [{"f": "fA"}, {"f": "fB"}]}),
-        );
-    }
+    assert_eq!(
+        serde_json::to_value(&response.data).unwrap(),
+        serde_json::json!({"us": [{"f": "fA"}, {"f": "fB"}]}),
+    );
+}
 
-    #[tokio::test]
-    async fn aliased_subgraph_data_rewrites_on_non_root_fetch() {
-        let schema = r#"
+#[tokio::test]
+async fn aliased_subgraph_data_rewrites_on_non_root_fetch() {
+    let schema = r#"
           schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
             @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
@@ -2775,7 +1931,7 @@ mod tests {
           }
         "#;
 
-        let query = r#"
+    let query = r#"
           {
             t {
               us {
@@ -2785,7 +1941,7 @@ mod tests {
           }
         "#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             ("S1", MockSubgraph::builder()
                 .with_json(
                     serde_json::json! {{
@@ -2821,32 +1977,32 @@ mod tests {
                 .build()),
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .query(query)
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .query(query)
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
-        let response = stream.next_response().await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
+    let response = stream.next_response().await.unwrap();
 
-        assert_eq!(
-            serde_json::to_value(&response.data).unwrap(),
-            serde_json::json!({"t": {"us": [{"f": "fA"}, {"f": "fB"}]}}),
-        );
-    }
+    assert_eq!(
+        serde_json::to_value(&response.data).unwrap(),
+        serde_json::json!({"t": {"us": [{"f": "fA"}, {"f": "fB"}]}}),
+    );
+}
 
-    #[tokio::test]
-    async fn errors_on_nullified_paths() {
-        let schema = r#"
+#[tokio::test]
+async fn errors_on_nullified_paths() {
+    let schema = r#"
           schema
             @link(url: "https://specs.apollo.dev/link/v1.0")
             @link(url: "https://specs.apollo.dev/join/v0.1", for: EXECUTION)
@@ -2900,7 +2056,7 @@ mod tests {
           }
         "#;
 
-        let query = r#"
+    let query = r#"
           query Query {
             foo {
               id
@@ -2912,7 +2068,7 @@ mod tests {
           }
         "#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
         ("S1", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"query Query__S1__0{foo{id bar{__typename id}}}", "operationName": "Query__S1__0"}},
                 serde_json::json!{{"data": {
@@ -2955,29 +2111,29 @@ mod tests {
         ).build())
     ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(query)
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(query)
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn missing_entities() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn missing_entities() {
+    let subgraphs = MockedSubgraphs([
             ("user", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"{currentUser{id activeOrganization{__typename id}}}"}},
                 serde_json::json!{{"data": {"currentUser": { "__typename": "User", "id": "0", "activeOrganization": { "__typename": "Organization", "id": "1" } } } }}
@@ -2986,29 +2142,29 @@ mod tests {
                                                        serde_json::json!{{"data": {}, "errors":[{"message":"error"}]}}).build())
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(SCHEMA)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(SCHEMA)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query("query { currentUser { id  activeOrganization{ id name } } }")
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query("query { currentUser { id  activeOrganization{ id name } } }")
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        insta::assert_json_snapshot!(stream.next_response().await.unwrap());
-    }
+    insta::assert_json_snapshot!(stream.next_response().await.unwrap());
+}
 
-    #[tokio::test]
-    async fn no_typename_on_interface() {
-        let subgraphs = MockedSubgraphs([
+#[tokio::test]
+async fn no_typename_on_interface() {
+    let subgraphs = MockedSubgraphs([
             ("animal", MockSubgraph::builder().with_json(
                 serde_json::json!{{"query":"query dog__animal__0{dog{id name}}", "operationName": "dog__animal__0"}},
                 serde_json::json!{{"data":{"dog":{"id":"4321","name":"Spot"}}}}
@@ -3021,7 +2177,7 @@ mod tests {
             ).build()),
         ].into_iter().collect());
 
-        let service = TestHarness::builder()
+    let service = TestHarness::builder()
             .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
             .unwrap()
             .schema(
@@ -3075,10 +2231,10 @@ mod tests {
             .await
             .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                "query dog {
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(
+            "query dog {
                 dog {
                   ...on Animal {
                     id
@@ -3088,19 +2244,19 @@ mod tests {
                   }
                 }
               }",
-            )
-            .build()
-            .unwrap();
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.clone().oneshot(request).await.unwrap();
+    let mut stream = service.clone().oneshot(request).await.unwrap();
 
-        let no_typename = stream.next_response().await.unwrap();
-        insta::assert_json_snapshot!(no_typename);
+    let no_typename = stream.next_response().await.unwrap();
+    insta::assert_json_snapshot!(no_typename);
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                "query dog {
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(
+            "query dog {
                 dog {
                   ...on Animal {
                     id
@@ -3111,40 +2267,40 @@ mod tests {
                   }
                 }
               }",
-            )
-            .build()
-            .unwrap();
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.clone().oneshot(request).await.unwrap();
+    let mut stream = service.clone().oneshot(request).await.unwrap();
 
-        let with_typename = stream.next_response().await.unwrap();
-        assert_eq!(
-            with_typename
-                .data
-                .clone()
-                .unwrap()
-                .get("dog")
-                .unwrap()
-                .get("name")
-                .unwrap(),
-            no_typename
-                .data
-                .clone()
-                .unwrap()
-                .get("dog")
-                .unwrap()
-                .get("name")
-                .unwrap(),
-            "{:?}\n{:?}",
-            with_typename,
-            no_typename
-        );
-        insta::assert_json_snapshot!(with_typename);
+    let with_typename = stream.next_response().await.unwrap();
+    assert_eq!(
+        with_typename
+            .data
+            .clone()
+            .unwrap()
+            .get("dog")
+            .unwrap()
+            .get("name")
+            .unwrap(),
+        no_typename
+            .data
+            .clone()
+            .unwrap()
+            .get("dog")
+            .unwrap()
+            .get("name")
+            .unwrap(),
+        "{:?}\n{:?}",
+        with_typename,
+        no_typename
+    );
+    insta::assert_json_snapshot!(with_typename);
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(
-                "query dog {
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(
+            "query dog {
                     dog {
                         ...on Dog {
                             name
@@ -3154,40 +2310,40 @@ mod tests {
                         }
                     }
                 }",
-            )
-            .build()
-            .unwrap();
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.oneshot(request).await.unwrap();
+    let mut stream = service.oneshot(request).await.unwrap();
 
-        let with_reversed_fragments = stream.next_response().await.unwrap();
-        assert_eq!(
-            with_reversed_fragments
-                .data
-                .clone()
-                .unwrap()
-                .get("dog")
-                .unwrap()
-                .get("name")
-                .unwrap(),
-            no_typename
-                .data
-                .clone()
-                .unwrap()
-                .get("dog")
-                .unwrap()
-                .get("name")
-                .unwrap(),
-            "{:?}\n{:?}",
-            with_reversed_fragments,
-            no_typename
-        );
-        insta::assert_json_snapshot!(with_reversed_fragments);
-    }
+    let with_reversed_fragments = stream.next_response().await.unwrap();
+    assert_eq!(
+        with_reversed_fragments
+            .data
+            .clone()
+            .unwrap()
+            .get("dog")
+            .unwrap()
+            .get("name")
+            .unwrap(),
+        no_typename
+            .data
+            .clone()
+            .unwrap()
+            .get("dog")
+            .unwrap()
+            .get("name")
+            .unwrap(),
+        "{:?}\n{:?}",
+        with_reversed_fragments,
+        no_typename
+    );
+    insta::assert_json_snapshot!(with_reversed_fragments);
+}
 
-    #[tokio::test]
-    async fn multiple_interface_types() {
-        let schema = r#"
+#[tokio::test]
+async fn multiple_interface_types() {
+    let schema = r#"
       schema
         @link(url: "https://specs.apollo.dev/link/v1.0")
         @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION) {
@@ -3316,7 +2472,7 @@ mod tests {
       }
       "#;
 
-        let query = r#"fragment OperationItemFragment on OperationItem {
+    let query = r#"fragment OperationItemFragment on OperationItem {
             __typename
             ... on OperationItemStuff {
               __typename
@@ -3353,7 +2509,7 @@ mod tests {
             }
           }"#;
 
-        let subgraphs = MockedSubgraphs([
+    let subgraphs = MockedSubgraphs([
             // The response isn't interesting to us,
             // we just need to make sure the query makes it through parsing and validation
             ("graph1", MockSubgraph::builder().with_json(
@@ -3362,83 +2518,82 @@ mod tests {
             ).build()),
             ].into_iter().collect());
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(schema)
-            .extra_plugin(subgraphs)
-            .build_supergraph()
-            .await
-            .unwrap();
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(schema)
+        .extra_plugin(subgraphs)
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let request = supergraph::Request::fake_builder()
-            .context(defer_context())
-            .query(query)
-            .variables(
-                serde_json_bytes::json! {{ "id": "1234", "a": 1, "b": 2}}
-                    .as_object()
-                    .unwrap()
-                    .clone(),
-            )
-            .build()
-            .unwrap();
+    let request = supergraph::Request::fake_builder()
+        .context(defer_context())
+        .query(query)
+        .variables(
+            serde_json_bytes::json! {{ "id": "1234", "a": 1, "b": 2}}
+                .as_object()
+                .unwrap()
+                .clone(),
+        )
+        .build()
+        .unwrap();
 
-        let mut stream = service.clone().oneshot(request).await.unwrap();
-        let response = stream.next_response().await.unwrap();
-        assert_eq!(serde_json_bytes::Value::Null, response.data.unwrap());
-    }
+    let mut stream = service.clone().oneshot(request).await.unwrap();
+    let response = stream.next_response().await.unwrap();
+    assert_eq!(serde_json_bytes::Value::Null, response.data.unwrap());
+}
 
-    #[tokio::test]
-    async fn id_scalar_can_overflow_i32() {
-        // Hack to let the first subgraph fetch contain an ID variable:
-        // ```
-        // type Query {
-        //     user(id: ID!): User @join__field(graph: USER)
-        // }
-        // ```
-        assert!(SCHEMA.contains("currentUser:"));
-        let schema = SCHEMA.replace("currentUser:", "user(id: ID!):");
+#[tokio::test]
+async fn id_scalar_can_overflow_i32() {
+    // Hack to let the first subgraph fetch contain an ID variable:
+    // ```
+    // type Query {
+    //     user(id: ID!): User @join__field(graph: USER)
+    // }
+    // ```
+    assert!(SCHEMA.contains("currentUser:"));
+    let schema = SCHEMA.replace("currentUser:", "user(id: ID!):");
 
-        let service = TestHarness::builder()
-            .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
-            .unwrap()
-            .schema(&schema)
-            .subgraph_hook(|_subgraph_name, _service| {
-                tower::service_fn(|request: subgraph::Request| async move {
-                    let id = &request.subgraph_request.body().variables["id"];
-                    Err(format!("$id = {id}").into())
-                })
-                .boxed()
+    let service = TestHarness::builder()
+        .configuration_json(serde_json::json!({"include_subgraph_errors": { "all": true } }))
+        .unwrap()
+        .schema(&schema)
+        .subgraph_hook(|_subgraph_name, _service| {
+            tower::service_fn(|request: subgraph::Request| async move {
+                let id = &request.subgraph_request.body().variables["id"];
+                Err(format!("$id = {id}").into())
             })
-            .build_supergraph()
-            .await
-            .unwrap();
+            .boxed()
+        })
+        .build_supergraph()
+        .await
+        .unwrap();
 
-        let large: i64 = 1 << 53;
-        let large_plus_one = large + 1;
-        // f64 rounds since it doesn’t have enough mantissa bits
-        assert!(large_plus_one as f64 as i64 == large);
-        // i64 of course doesn’t round
-        assert!(large_plus_one != large);
+    let large: i64 = 1 << 53;
+    let large_plus_one = large + 1;
+    // f64 rounds since it doesn’t have enough mantissa bits
+    assert!(large_plus_one as f64 as i64 == large);
+    // i64 of course doesn’t round
+    assert!(large_plus_one != large);
 
-        let request = supergraph::Request::fake_builder()
-            .query("query($id: ID!) { user(id: $id) { name }}")
-            .variable("id", large_plus_one)
-            .build()
-            .unwrap();
-        let response = service
-            .oneshot(request)
-            .await
-            .unwrap()
-            .next_response()
-            .await
-            .unwrap();
-        // The router did not panic or respond with an early validation error.
-        // Instead it did a subgraph fetch, which recieved the correct ID variable without rounding:
-        assert_eq!(
-            response.errors[0].extensions["reason"].as_str().unwrap(),
-            "$id = 9007199254740993"
-        );
-        assert_eq!(large_plus_one.to_string(), "9007199254740993");
-    }
+    let request = supergraph::Request::fake_builder()
+        .query("query($id: ID!) { user(id: $id) { name }}")
+        .variable("id", large_plus_one)
+        .build()
+        .unwrap();
+    let response = service
+        .oneshot(request)
+        .await
+        .unwrap()
+        .next_response()
+        .await
+        .unwrap();
+    // The router did not panic or respond with an early validation error.
+    // Instead it did a subgraph fetch, which recieved the correct ID variable without rounding:
+    assert_eq!(
+        response.errors[0].extensions["reason"].as_str().unwrap(),
+        "$id = 9007199254740993"
+    );
+    assert_eq!(large_plus_one.to_string(), "9007199254740993");
 }
